+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Summoned entities no longer suffocate
 - Improve summoned entity placements, to avoid placing them into solid blocks
+- Add spell modifier `teleport_distance_add` field
 
 # 1.9.12